Unify securityHandler testing across versions
