+++ examples/generated-javascript-project/package.json
@@ -16,7 +16,7 @@
 		"@seriousme/openapi-schema-validator": "^2.2.1",
 		"js-yaml": "^4.1.0",
 		"minimist": "^1.2.8",
-		"fastify-openapi-glue": "^4.5.0"
+		"fastify-openapi-glue": "^4.6.0"
 	},
 	"devDependencies": {
 		"fastify": "^4.26.2",
