+++ test/test-plugin.v2.js
@@ -216,7 +216,7 @@ test("invalid service definition throws error ", (t, done) => {
 test("full pet store V2 definition does not throw error ", (t, done) => {
 	const fastify = Fastify(noStrict);
 	fastify.register(fastifyOpenapiGlue, {
-		specification: JSON.parse(JSON.stringify(petStoreSpec)),
+		specification: structuredClone(petStoreSpec),
 		serviceHandlers,
 	});
 	fastify.ready((err) => {
@@ -294,7 +294,7 @@ test("generic path parameters override works", async (t) => {
 test("schema attributes for non-body parameters work", async (t) => {
 	const fastify = Fastify(noStrict);
 	fastify.register(fastifyOpenapiGlue, {
-		specification: JSON.parse(JSON.stringify(petStoreSpec)),
+		specification: structuredClone(petStoreSpec),
 		serviceHandlers,
 	});
 	const res = await fastify.inject({
